move cleanup into util mod
